// src/types/mapsInterface.ts
export interface ApiMap {
    uuid: string;
    displayName: string;
    tacticalDescription?: string;
    coordinates?: string;
    displayIcon?: string;
    listViewIcon?: string;
    listViewIconTall?: string;
    splash?: string;
    stylizedBackgroundImage?: string;
    premierBackgroundImage?: string;
  }
  